refactor: triplestore errors
