+++ webapi/src/main/scala/org/knora/webapi/http/status/ApiStatusCodesV2.scala
@@ -8,6 +8,7 @@ package org.knora.webapi.http.status
 import akka.http.scaladsl.model.StatusCode
 import akka.http.scaladsl.model.StatusCodes
 import dsp.errors._
+import org.knora.webapi.store.triplestore.errors.TriplestoreTimeoutException
 
 /**
  * The possible values for the HTTP status code that is returned as part of each Knora API v2 response.
